Remove component for fetching layer groups; rename annotations support;
