--- app/templates/components/project-geometries/layer-groups.hbs
@@ -1,7 +1,0 @@
-{{#if model}}
-  {{yield (hash
-    model=model
-    projectAreaSource=projectAreaSource
-    projectBufferSource=projectBufferSource
-    handleMapLoad=handleMapLoad)}}
-{{/if}}
