+++ app/utils/mapbox-gl-draw/annotations/generate-curvature.js
