+++ app/utils/mapbox-gl-draw/annotations/styles.js
