--- tests/integration/components/project-geometries/layer-groups-test.js
@@ -1,18 +1,0 @@
-import { module, test } from 'qunit';
-import { setupRenderingTest } from 'ember-qunit';
-import { render } from '@ember/test-helpers';
-import hbs from 'htmlbars-inline-precompile';
-import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
-
-module('Integration | Component | project-geometries/layer-groups', function(hooks) {
-  setupRenderingTest(hooks);
-  setupMirage(hooks);
-
-  test('it renders', async function(assert) {
-    this.server.createList('layer-group', 10);
-
-    await render(hbs`{{project-geometries/layer-groups}}`);
-
-    assert.ok(this);
-  });
-});
