--- tests/unit/utils/mapbox-gl-draw/annotation-generate-curvature-test.js
@@ -1,11 +1,0 @@
-import mapboxGlDrawAnnotationGenerateCurvature from 'labs-applicant-maps/utils/mapbox-gl-draw/annotation-generate-curvature';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | mapbox-gl-draw/annotation-generate-curvature', function(hooks) {
-
-  // Replace this with your real tests.
-  test('it works', function(assert) {
-    let result = mapboxGlDrawAnnotationGenerateCurvature();
-    assert.ok(result);
-  });
-});
