--- tests/unit/utils/mapbox-gl-draw/annotation-mode-test.js
@@ -1,11 +1,0 @@
-import mapboxGlDrawAnnotationMode from 'labs-applicant-maps/utils/mapbox-gl-draw/annotation-mode';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | mapbox-gl-draw/annotation-mode', function(hooks) {
-
-  // Replace this with your real tests.
-  test('it works', function(assert) {
-    let result = mapboxGlDrawAnnotationMode();
-    assert.ok(result);
-  });
-});
