--- tests/unit/utils/mapbox-gl-draw/annotation-styles-test.js
@@ -1,11 +1,0 @@
-import mapboxGlDrawAnnotationStyles from 'labs-applicant-maps/utils/mapbox-gl-draw/annotation-styles';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | mapbox-gl-draw/annotation-styles', function(hooks) {
-
-  // Replace this with your real tests.
-  test('it works', function(assert) {
-    let result = mapboxGlDrawAnnotationStyles();
-    assert.ok(result);
-  });
-});
