+++ tests/unit/utils/mapbox-gl-draw/annotations/generate-curvature-test.js
@@ -1,0 +1,10 @@
+import mapboxGlDrawAnnotationGenerateCurvature from 'labs-applicant-maps/utils/mapbox-gl-draw/annotations/generate-curvature';
+import { module, skip } from 'qunit';
+
+module('Unit | Utility | mapbox-gl-draw/annotation-generate-curvature', function() {
+  // Replace this with your real tests.
+  skip('it works', function(assert) {
+    const result = mapboxGlDrawAnnotationGenerateCurvature();
+    assert.ok(result);
+  });
+});
