+++ tests/unit/utils/mapbox-gl-draw/annotations/mode-test.js
@@ -1,0 +1,10 @@
+import mapboxGlDrawAnnotationMode from 'labs-applicant-maps/utils/mapbox-gl-draw/annotations/mode';
+import { module, skip } from 'qunit';
+
+module('Unit | Utility | mapbox-gl-draw/annotation-mode', function() {
+  // Replace this with your real tests.
+  skip('it works', function(assert) {
+    const result = mapboxGlDrawAnnotationMode();
+    assert.ok(result);
+  });
+});
