+++ tests/unit/utils/mapbox-gl-draw/annotations/styles-test.js
@@ -1,0 +1,10 @@
+import mapboxGlDrawAnnotationStyles from 'labs-applicant-maps/utils/mapbox-gl-draw/annotations/styles';
+import { module, skip } from 'qunit';
+
+module('Unit | Utility | mapbox-gl-draw/annotation-styles', function() {
+  // Replace this with your real tests.
+  skip('it works', function(assert) {
+    const result = mapboxGlDrawAnnotationStyles();
+    assert.ok(result);
+  });
+});
